--assume-new implies all other files are old
This means that the --trace will only print targets updated due to the
listed --assume-new files being new, and ignore the state of all other
files. It also avoid stat() calls for all files if --assume-new is given
and --assume-old is not.

If both assume forms are used (new and old), then the unnamed files will
be stat()'ed and the result outcome will reflect them.

This allows checking if a target depends on a source:

make -q --assume-new=source.c target.log returns 1 if target.log depends
on source.c and 0 otherwise, whether source.c exists or not.